pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verdict"
version = "0.1.0"
description = "Attribution methods for medical vision-language models - VERDICT"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Anish", email = "your.email@imperial.ac.uk"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
keywords = [
    "medical imaging",
    "vision-language models",
    "attribution",
    "interpretability",
    "chest x-ray",
    "radiology",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "bitsandbytes>=0.41.0",
    "captum>=0.6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pydicom>=2.4.0",
    "SimpleITK>=2.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "opencv-python>=4.8.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "wandb>=0.15.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
    "ipython>=8.12.0",
    "notebook>=7.0.0",
    "ipywidgets>=8.1.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/verdict"
Documentation = "https://github.com/yourusername/verdict#readme"
Repository = "https://github.com/yourusername/verdict"
Issues = "https://github.com/yourusername/verdict/issues"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*", "notebooks"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | outputs
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["verdict", "data", "models", "attribution", "baselines", "evaluation", "experiments", "utils"]
skip = [".git", ".venv", "outputs"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "outputs/",
    "notebooks/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["data", "models", "attribution", "baselines", "evaluation", "experiments", "utils"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
